refactor(post): add like and unlike functionality
Added `handlePostLike` and `handlePostUnlike` functions to the `postController` module to handle liking and unliking a post. These functions update the `likedBy` array of a post document in the database based on the user's action. The routes for liking and unliking a post have also been added to the `post.js` file in the `server/routes` directory.